+++ CHANGELOG.md
@@ -1,4 +1,21 @@
 Change log
 
+# 1.2.1 (14.05.2020)
+
+* Synced with ZombieBox 2.7
+* Fixed import paths in templates
+
+# 1.2.0 (03.06.2019)
+
+* Synced with ZombieBox 2.0.0 and CuteJS 2.0.0
+
+# 1.1.1 (23.05.2019)
+
+* Fixed issue when extension was not correctly listening to source file changes
+
+# 1.1.0 (26.04.2019)
+
+Synchronise with ZombieBox 2.0.0-alpha.7
+
 # 1.0.0 (06.03.2019)
 Initial release